* 'develope' of https://github.com/Practical-Project-10/SkiFriend-BackEnd:
  Carpool 게시물 유효성 검사를 위한 Enum
  카풀 회원만 수정 삭제
  전체 조회
  카풀 Dto에 Status값 내려주기
  카풀 게시글 모집완료 기능, 내가 쓴 게시물 목록 불러오기 기능 완료
  전체 조회
  카풀 게시글 모집완료 여부 상태 표시 컬럼 추가

+++ src/main/java/com/ppjt10/skifriend/service/CarpoolService.java
@@ -23,18 +23,37 @@ public class CarpoolService {
     }
 
     @Transactional
-    public void updateCarpool(Long carpoolId, CarpoolDto.RequestDto requestDto) {
+    public void updateCarpool(Long carpoolId, CarpoolDto.RequestDto requestDto, Long userid) {
         Carpool carpool = carpoolRepository.findById(carpoolId).orElseThrow(
                 () -> new IllegalArgumentException("해당 아이디의 카풀이 존재하지 않습니다.")
         );
+        if(carpool.getUser().getId() != userid){
+            throw new IllegalArgumentException("작성자만 상태를 변경할 수 있습니다.");
+        }
         carpool.update(requestDto);
     }
 
     @Transactional
-    public void deleteCarpool(Long carpoolId) {
+    public void deleteCarpool(Long carpoolId, Long userid) {
         Carpool carpool = carpoolRepository.findById(carpoolId).orElseThrow(
                 () -> new IllegalArgumentException("해당 아이디의 카풀이 존재하지 않습니다.")
         );
+        if(carpool.getUser().getId() != userid){
+            throw new IllegalArgumentException("작성자만 상태를 변경할 수 있습니다.");
+        }
         carpoolRepository.deleteById(carpoolId);
+    }
+
+    @Transactional
+    public void changeStatus(Long carpoolId, Long userid) {
+        Carpool carpool = carpoolRepository.findById(carpoolId).orElseThrow(
+                () -> new IllegalArgumentException("해당 아이디의 카풀이 존재하지 않습니다.")
+        );
+
+        if(carpool.getUser().getId() != userid){
+            throw new IllegalArgumentException("작성자만 상태를 변경할 수 있습니다.");
+        }
+
+        carpool.changeStatus();
     }
 }
